# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Evolution GO - whatsmeow",
        "title": "Evolution GO",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/newsletter/create": {
            "post": {
                "description": "Create newsletter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Newsletter"
                ],
                "summary": "Create newsletter",
                "parameters": [
                    {
                        "description": "Newsletter data",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Zapbox-API_evolution-go_pkg_newsletter_service.CreateNewsletterStruct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Error on validation",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/newsletter/info": {
            "post": {
                "description": "Get newsletter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Newsletter"
                ],
                "summary": "Get newsletter",
                "parameters": [
                    {
                        "description": "Newsletter data",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Zapbox-API_evolution-go_pkg_newsletter_service.GetNewsletterStruct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Error on validation",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/newsletter/link": {
            "post": {
                "description": "Get newsletter invite",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Newsletter"
                ],
                "summary": "Get newsletter invite",
                "parameters": [
                    {
                        "description": "Newsletter data",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Zapbox-API_evolution-go_pkg_newsletter_service.GetNewsletterInviteStruct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Error on validation",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/newsletter/list": {
            "get": {
                "description": "List newsletters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Newsletter"
                ],
                "summary": "List newsletters",
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/newsletter/messages": {
            "post": {
                "description": "Get newsletter messages",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Newsletter"
                ],
                "summary": "Get newsletter messages",
                "parameters": [
                    {
                        "description": "Newsletter data",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Zapbox-API_evolution-go_pkg_newsletter_service.GetNewsletterMessagesStruct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Error on validation",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/newsletter/subscribe": {
            "post": {
                "description": "Subscribe newsletter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Newsletter"
                ],
                "summary": "Subscribe newsletter",
                "parameters": [
                    {
                        "description": "Newsletter data",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Zapbox-API_evolution-go_pkg_newsletter_service.GetNewsletterStruct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Error on validation",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "gin.H": {
            "type": "object",
            "additionalProperties": {}
        },
...
        "github_com_Zapbox-API_evolution-go_pkg_newsletter_service.CreateNewsletterStruct": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_Zapbox-API_evolution-go_pkg_newsletter_service.GetNewsletterInviteStruct": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                }
            }
        },
        "github_com_Zapbox-API_evolution-go_pkg_newsletter_service.GetNewsletterMessagesStruct": {
            "type": "object",
            "properties": {
                "before_id": {
                    "type": "integer"
                },
                "count": {
                    "type": "integer"
                },
                "jid": {
                    "$ref": "#/definitions/types.JID"
                }
            }
        },
        "github_com_Zapbox-API_evolution-go_pkg_newsletter_service.GetNewsletterStruct": {
            "type": "object",
            "properties": {
                "jid": {
                    "$ref": "#/definitions/types.JID"
                }
            }
        },
...
        "types.JID": {
            "type": "object",
            "properties": {
                "device": {
                    "type": "integer"
                },
                "integrator": {
                    "type": "integer"
                },
                "rawAgent": {
                    "type": "integer"
                },
                "server": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                }
            }
        },
...
    }
}